### 2023-03-10 a8edc80
Fix link to newcomers' guide (#232)
