### 2023-03-27 #5242 @xuanxu
@j-faria would you be willing to review this paper? I think your expertise with astrophysical Pythonmakes you a nice fit for this submission.
