### 2023-03-31 #5318 @editorialbot
[REVIEW]: sunraster: Manipulating and Visualizing Solar Slit-spectrograph Observations in Python
**Submitting author:** @DanRyanIrish (<a href="http://orcid.org/0000-0001-8661-3825">Daniel Ryan</a>)
**Repository:** https://github.com/sunpy/sunraster
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.3
