+++ .pre-commit-config.yaml
@@ -1,9 +1,4 @@
 repos:
-  - repo: https://github.com/PyCQA/docformatter
-    rev: v1.6.0
-    hooks:
-      - id: docformatter
-        args: [--in-place, --pre-summary-newline, --make-summary-multi]
   - repo: https://github.com/PyCQA/autoflake
     rev: v2.0.2
     hooks:
