### 2023-04-04 a6d4114
[pre-commit.ci] pre-commit autoupdate (#235)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Nabil Freij <nabil.freij@gmail.com>
