+++ .pre-commit-config.yaml
@@ -1,9 +1,4 @@
 repos:
-  - repo: https://github.com/PyCQA/docformatter
-    rev: v1.5.1
-    hooks:
-      - id: docformatter
-        args: [--in-place, --pre-summary-newline, --make-summary-multi]
   - repo: https://github.com/PyCQA/autoflake
     rev: v2.0.2
     hooks:
@@ -11,7 +6,7 @@ repos:
         args: ['--in-place', '--remove-all-unused-imports', '--remove-unused-variable']
         exclude: ".*(.fits|.fts|.fit|.txt|tca.*|extern.*|.rst|.md|__init__.py|docs/conf.py)$"
   -   repo: https://github.com/psf/black
-      rev: 23.1.0
+      rev: 23.3.0
       hooks:
       - id: black
         exclude: ".*(.fits|.fts|.fit|.txt|.csv)$"
