### 2023-04-26 #5318 @j-faria
Hi @DanRyanIrish, I've now started the review of `sunraster` (apologies for the delay).
I have installed the package and went through the examples in the documentation; everything worked perfectly. In order to tick the remaining boxes, I have only one simple request and a couple of questions:
- I'm missing from the documentation an example of using `sunraster` on a real-world analysis.
  Related to this, would it be possible to include with the package (or automatically fetch from somewhere else) an example dataset that is typical of what `sunraster` will be used for?
  Because this example is missing, I think the documentation also misses a figure or two. For example, the plot generated [in this section](https://docs.sunpy.org/projects/sunraster/en/latest/data_types/spectrogram.html#plotting) could be included in the page? Admittedly, the plot is not very interesting because the data is just ones, but an example dataset would be more interesting.

- I'm assuming the package is being tested continuously with [this action](https://github.com/sunpy/sunraster/actions/workflows/ci.yml), can you please confirm?
- Are there other packages that are commonly used to achieve the same analysis that `sunraster` allows for?
