### 2023-04-27 #5318 @DanRyanIrish
Hi @j-faria.  Thanks so much for your review and for the positive feedback.

>     * I'm missing from the documentation an example of using `sunraster` on a real-world analysis.
>       Related to this, would it be possible to include with the package (or automatically fetch from somewhere else) an example dataset that is typical of what `sunraster` will be used for?
>       Because this example is missing, I think the documentation also misses a figure or two. For example, the plot generated [in this section](https://docs.sunpy.org/projects/sunraster/en/latest/data_types/spectrogram.html#plotting) could be included in the page? Admittedly, the plot is not very interesting because the data is just ones, but an example dataset would be more interesting.

Thanks for this helpful suggestion.  I agree that this would be a helpful addition for users.  I will give some thought as to the best way to implement it and let you know when it is ready to be viewed.

>     * I'm assuming the package is being tested continuously with [this action](https://github.com/sunpy/sunraster/actions/workflows/ci.yml), can you please confirm?

Yes that's right.

>     * Are there other packages that are commonly used to achieve the same analysis that `sunraster` allows for?

Regarding other packages that work with `sunraster`, much of the underlying infrastructure comes from the `ndcube` package.  Updates in that package transparently become available through the `sunraster` data classes.  For example, rebinning data into "macropixels", e.g. turning a 8x8 grid into a 4x4 grid by summing/averaging neighbouring 2x2 pixel areas, recently became available in `ndcube`.  This functionality can now used used directly through the `sunraster` data classes and is often used in solar slit spectroscopy to boost signal-to-noise before performing spectral fits.  Moreover, because the `sunraster` objects are compatible with the ndcube APIs, tools built on top of those APIs should work seamlessly with the `sunraster` data classes.  Was this the question you were asking?
