+++ setup.cfg
@@ -41,7 +41,7 @@ install_requires =
 
 [options.extras_require]
 instr =
-  sunpy>=4.0.0
+  sunpy==5.0.0rc1
 tests =
   pytest-astropy
 docs =
