### 2023-05-18 sunpy/sunraster#241 @nabobalis
I think is is a result of picking the default unit as metres instead of nm.

I can work out how to fix that.
