### 2023-05-19 sunpy/sunraster#241 @Cadair
The default unit is unfortunately picked by `astropy.wcs.WCS`, which defaults everything to SI. What we need to do is write a unit-preserving `WCS` wrapper in astropy core, because by the time NDCube / WCSAxes get's given the `WCS` object there's no indication that it used to be in nm left anywhere.
