### 2023-05-26 #5318 @DanRyanIrish
Thanks so much @mwcraig for this review.  I will consider how to best address these comments.  In the meantime, here are a few initial responses.

>
>     * Several of the builtin plots [looked odd when I rendered them](https://github.com/sunpy/sunraster/issues/241)

As highlighted on the issue thread, the root issue is that the `WCS` infrastructure in the astropy core package does not preserve the coordinate units but defaults everything to SI.  However, perhaps there's a way of making the example code and built-in plots manually override this.  This will make the docs nicer but won't prevent plots like this being produced by users in real-world analysis.  But is might help to show users how to overcome this issue themselves.

> and [one failed](https://github.com/sunpy/sunraster/issues/240).

This is an issue caused by changes to the latest releases of `ndcube` and needs addressing by `sunraster`.  Thanks for spotting it.

>     * All of the plots would benefit from the change that @j-faria suggested: add a more real-world example. Even if adding a truly realistic example isn't feasible because of the size of the data, amore interesting toy example would better demonstrate what `sunraster` accomplishes. Both the documentation and the paper would be improved with this.

Thanks for highlighting the importance of this addition.

>     * The documentation is, overall, thorough. A few improvements are needed, though:

This feedback is very helpful as sometimes it requires a fresh pair of eyes to spot what's not clear.

>     * The paper didn't discuss whether there are other packages that provide similar functionality. My guess is that there are not, but a statement to that effect would be helpful.

There is no other Python package that specifically addresses the needs of handling data from solar slit spectrometers.  This fact can be added to the paper.
