+++ tox.ini
@@ -11,7 +11,6 @@ requires =
 
 [testenv]
 pypi_filter = https://raw.githubusercontent.com/sunpy/sunpy/main/.test_package_pins.txt
-passenv = HOME WINDIR LC_ALL LC_CTYPE CC CI TRAVIS
 changedir = .tmp/{envname}
 description =
     run tests
@@ -27,7 +26,7 @@ deps =
     devdeps: git+https://github.com/astropy/astropy
     devdeps: git+https://github.com/sunpy/ndcube
     devdeps: git+https://github.com/sunpy/sunpy
-    oldestdeps: ndcube<3.0.0
+    oldestdeps: ndcube<2.2
     online: pytest-rerunfailures
     online: pytest-timeout
     pytest-cov
