+++ setup.cfg
@@ -14,14 +14,13 @@ github_project = sunpy/sunraster
 platform = any
 keywords = solar physics, solar, science, sun, wcs, coordinates, spectra, raster
 classifiers =
-  Development Status :: 5 - Production/Stable
+  Development Status :: 4 - Beta
   Intended Audience :: Science/Research
   License :: OSI Approved :: BSD License
   Natural Language :: English
   Operating System :: OS Independent
   Programming Language :: Python
   Programming Language :: Python :: 3
-  Programming Language :: Python :: 3.8
   Programming Language :: Python :: 3.9
   Programming Language :: Python :: 3.10
   Programming Language :: Python :: 3.11
@@ -29,7 +28,7 @@ classifiers =
 
 [options]
 zip_safe = False
-python_requires = >=3.8
+python_requires = >=3.9
 packages = find:
 include_package_data = True
 setup_requires =
@@ -37,11 +36,11 @@ setup_requires =
 install_requires =
   numpy
   astropy
-  ndcube[plotting]>=2.0.1
+  ndcube[plotting]>=2.1.2
 
 [options.extras_require]
 instr =
-  sunpy>=4.0.0
+  sunpy>=5.0.0rc1
 tests =
   pytest-astropy
 docs =
@@ -50,8 +49,6 @@ docs =
   sphinx-changelog
   sphinx-gallery
   sunpy-sphinx-theme
-  # Not a direct dependency, but we need to pin this until sphinx-changelog is updated (see #6668)
-  towncrier<22.12.0
 
 [options.packages.find]
 exclude = sunraster._dev
