### 2023-07-08 #5318 @DanRyanIrish
Here I compile a list of action items emanating from the two reviews:
## Paper Improvements

- [ ] Mention that sunraster is unique in meeting the needs of solar slit spectrograph data analysis in Python.
- [ ] Add a plot of solar slit spectrograph data being manipulated by sunraster.

## Documentation Improvements

- [ ] Add a real world example in the docs, including plots.
- [ ] Set plots in docs to have readable (non-SI) ticklabels.
- [ ] Add link to code of conduct
- [ ] Add "How to Contribute" section.  Include info on how to install testing envs and run tests.
