### 2023-07-08 d28183b
Add mention in JOSS paper that no other Python package provides the features of sunraster.
