### 2023-07-10 dfd6548
Remove unneeded parts of tox.ini
