+++ tox.ini
@@ -1,13 +1,9 @@
 [tox]
+min_version = 4.0
 envlist =
-    py{38,39,310,311}{,-oldestdeps,-devdeps,-online}
+    py{39,310,311}{,-oldestdeps,-devdeps,-online}
     build_docs
     codestyle
-isolated_build = true
-requires =
-    setuptools >=56, !=61.0.0
-    pip >= 19.3.1
-    tox-pypi-filter >= 0.12
 
 [testenv]
 pypi_filter = https://raw.githubusercontent.com/sunpy/sunpy/main/.test_package_pins.txt
