### 2023-07-10 b6a54ca
Remove unneeded parts of tox.ini (#246)
