+++ .pre-commit-config.yaml
@@ -16,7 +16,7 @@ repos:
       - id: isort
         exclude: ".*(.fits|.fts|.fit|.txt|.csv)$"
   -   repo: https://github.com/pre-commit/pre-commit-hooks
-      rev: v4.4.0
+      rev: v4.5.0
       hooks:
       - id: check-ast
       - id: check-case-conflict
