+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-0.5.0 (2023-06-06)
+0.5.0 (2023-11-16)
 ==================
 
 Breaking Changes
