### 2023-12-04 #5318 @dfm
@DanRyanIrish — I'm sorry to hear that this is the outcome here!

@xuanxu, @mwcraig, @j-faria — Thank you all for the time that you committed to this review! JOSS runs on volunteer effort and we couldn't do this without all of you and it's very much appreciated!!
